LayoutTests/storage/indexeddb/createObjectStore-null-name-expected.txt

Test IndexedDB createObjectStore null handling

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.open(name)
db = event.target.result
request = db.setVersion('1')
Deleted all object stores.
db.createObjectStore(null);
PASS objectStore.name is "null"
PASS successfullyParsed is true

TEST COMPLETE
